Interface work, package restructuring
I implemented most of the gamejam game interface in TicTacToe, it still
needs save/load features. It became clear that the MVC packages were
going to start getting cluttered so I
restructured the packages to have an overall MVC for non game-specific
stuff, and a specific game package for each game.